+++ silkrpc/coro/task.hpp
@@ -18,14 +18,18 @@
 #define SILKRPC_CORO_TASK_HPP
 
 #include <coroutine>
-#include <iostream>
+#include <functional>
 #include <thread>
+#include <variant>
+
+namespace silkrpc::coro {
 
 template<typename T>
 struct task {
     struct promise_type {
         std::variant<std::monostate, T, std::exception_ptr> result;
         std::coroutine_handle<> caller_coro;
+        std::function<void()> completion_handler;
 
         task get_return_object() { return {this}; }
         auto initial_suspend() { return std::suspend_always{}; }
@@ -34,9 +38,10 @@ struct task {
                 promise_type* self;
                 bool await_ready() { return false; }
                 void await_suspend(std::coroutine_handle<>) {
-                    std::cout << "task::promise_type::final_suspend::Awaiter::await_suspend thread=" << std::this_thread::get_id() << "\n" << std::flush;
                     if (self->caller_coro) {
                         self->caller_coro.resume();
+                    } else if (self->completion_handler) {
+                        self->completion_handler();
                     }
                 }
                 void await_resume() {}
@@ -46,7 +51,6 @@ struct task {
         template <typename U>
         void return_value(U&& value) {
             result.template emplace<1>(std::forward<U>(value));
-            std::cout << "task::return_value thread=" << std::this_thread::get_id() << " value=" << value << "\n" << std::flush;
         }
         void set_exception(std::exception_ptr eptr) {
             result.template emplace<2>(std::move(eptr));
@@ -66,7 +70,6 @@ struct task {
     bool await_ready() { return false; }
 
     auto await_suspend(std::coroutine_handle<> caller_coro) {
-        std::cout << "task::await_suspend thread=" << std::this_thread::get_id() << " caller_coro=" << caller_coro.address() << "\n" << std::flush;
         coro.promise().caller_coro = caller_coro;
         coro.resume();
     }
@@ -78,12 +81,13 @@ struct task {
         return std::get<1>(coro.promise().result);
     }
 
-    auto coroutine_handle() const { return coro; }
+    void start(std::function<void()> completion_handler = {}) {
+        coro.promise().completion_handler = completion_handler;
+        coro.resume();
+    }
 
 private:
-    task(promise_type* p) noexcept : coro(std::coroutine_handle<promise_type>::from_promise(*p)) {
-        std::cout << "task thread=" << std::this_thread::get_id() << " coro=" << coro.address() << "\n" << std::flush;
-    }
+    task(promise_type* p) noexcept : coro(std::coroutine_handle<promise_type>::from_promise(*p)) {}
 
     std::coroutine_handle<promise_type> coro;
 };
@@ -92,27 +96,26 @@ template<>
 struct task<void> {
     struct promise_type {
         std::coroutine_handle<> caller_coro;
+        std::function<void()> completion_handler;
 
-        task get_return_object() { return task{this}; }
+        task get_return_object() { return {this}; }
         auto initial_suspend() { return std::suspend_always{}; }
         auto final_suspend() {
             struct Awaiter {
                 promise_type* self;
                 bool await_ready() { return false; }
                 void await_suspend(std::coroutine_handle<>) {
-                    std::cout << "task::promise_type::final_suspend::Awaiter::await_suspend caller_coro=" << self->caller_coro.address() << "\n" << std::flush;
                     if (self->caller_coro) {
-                        std::cout << "task::promise_type::final_suspend::Awaiter::await_suspend thread=" << std::this_thread::get_id() << "\n" << std::flush;
                         self->caller_coro.resume();
+                    } else if (self->completion_handler) {
+                        self->completion_handler();
                     }
                 }
                 void await_resume() {}
             };
             return Awaiter{this};
         }
-        void return_void() {
-            std::cout << "task::return_void thread=" << std::this_thread::get_id() << "\n" << std::flush;
-        }
+        void return_void() {}
         void unhandled_exception() noexcept { std::terminate(); }
     };
 
@@ -122,29 +125,29 @@ struct task<void> {
     task& operator=(task&&) = delete;
 
     ~task() {
-        std::cout << "~task START thread=" << std::this_thread::get_id() << " coro=" << coro.address() << "\n" << std::flush;
         if (coro) coro.destroy();
-        std::cout << "~task END thread=" << std::this_thread::get_id() << " coro=" << coro.address() << "\n" << std::flush;
     }
 
     bool await_ready() { return false; }
 
     auto await_suspend(std::coroutine_handle<> caller_coro) {
-        std::cout << "task::await_suspend thread=" << std::this_thread::get_id() << " caller_coro=" << caller_coro.address() << "\n" << std::flush;
         coro.promise().caller_coro = caller_coro;
         coro.resume();
     }
 
     void await_resume() {}
 
-    auto coroutine_handle() const { return coro; }
+    void start(std::function<void()> completion_handler = {}) {
+        coro.promise().completion_handler = completion_handler;
+        coro.resume();
+    }
 
 private:
-    task(promise_type* p) : coro(std::coroutine_handle<promise_type>::from_promise(*p)) {
-        std::cout << "task thread=" << std::this_thread::get_id() << " coro=" << coro.address() << "\n" << std::flush;
-    }
+    task(promise_type* p) : coro(std::coroutine_handle<promise_type>::from_promise(*p)) {}
 
     std::coroutine_handle<promise_type> coro;
 };
+
+} // namespace silkrpc::coro
 
 #endif // SILKRPC_CORO_TASK_HPP
